Move ws-specific stuff to its own package
